Refactor full auto system into modular package
